swarm/pss: Omit build of tests using t.Name for go 1.7
